pyproject.toml
[project]
name = "portfolio"
version = "0.1.0"
description = "A modern, forkable portfolio website built with FastAPI and Jinja2"
authors = [
    {name = "Pedro Spinosa", email = "spinosaphb@gmail.com"},
]
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.1",
    "pydantic>=2.0.0",
]

requires-python = ">=3.11"
readme = "README.md"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]
dev = [
    "pre-commit>=3.0.0",
    "mypy>=1.0.0",
    "ruff>=0.4.0",
    "types-pyyaml>=6.0.0",
]

[project.scripts]
portfolio = "src.main:app"
start = "src.main:main"
dev = "src.main:dev"
test = "pytest:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["B011"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "run_tests"
disallow_untyped_defs = false
